requirements.txt
Pillow
django
dnspython